💚 updated documentation and SpgxAppCLIBulkAction.ts
